requirements.txt
mcp==1.0.0
tradingview-ta
google-cloud-aiplatform
google-cloud-firestore
google-cloud-storage
streamlit
